fix: some demo fixes
