+++ sessions/js-structure/demo-end/index.html
@@ -10,33 +10,40 @@
 	<body>
 		<header class="header">
 			<h1 data-js="headline">JS Structure</h1>
-			<button class="header__button" type="button" data-js="header-button">Click me!</button>
+			<button class="header__button" type="button" data-js="header-button">
+				Change Headline
+			</button>
 		</header>
 		<main>
-			<article class="card" data-js="card">
-				<h2>A typical Card</h2>
-				<p class="card__text">
-					Lorem, ipsum dolor sit amet consectetur adipisicing elit. Porro veniam, mollitia
-					facilis officiis nam saepe tenetur recusandae rem corrupti labore id, delectus
-					quas quos quasi expedita pariatur esse, necessitatibus assumenda.
-				</p>
-				<button class="card__button" type="button" data-js="card-button">
-					Rephrase Card
-				</button>
-			</article>
-			<article class="card" data-js="card">
-				<h2>Another typical Card</h2>
-				<p class="card__text">
-					Lorem 2, ipsum dolor sit amet consectetur adipisicing elit. Porro veniam,
-					mollitia facilis officiis nam saepe tenetur recusandae rem corrupti labore id,
-					delectus quas quos quasi expedita pariatur esse, necessitatibus assumenda.
-				</p>
-				<button class="card__button" type="button" data-js="card-button">
-					Rephrase Card
-				</button>
-			</article>
+			<section class="cards">
+				<article class="card" data-js="card">
+					<h2>A typical Card</h2>
+					<p class="card__text">
+						Lorem, ipsum dolor sit amet consectetur adipisicing elit. Porro veniam,
+						mollitia facilis officiis nam saepe tenetur recusandae rem corrupti labore
+						id, delectus quas quos quasi expedita pariatur esse, necessitatibus
+						assumenda.
+					</p>
+					<button class="card__button" type="button" data-js="card-button">
+						Rephrase Card
+					</button>
+				</article>
+				<article class="card" data-js="card">
+					<h2>Another typical Card</h2>
+					<p class="card__text">
+						Lorem 2, ipsum dolor sit amet consectetur adipisicing elit. Porro veniam,
+						mollitia facilis officiis nam saepe tenetur recusandae rem corrupti labore
+						id, delectus quas quos quasi expedita pariatur esse, necessitatibus
+						assumenda.
+					</p>
+					<button class="card__button" type="button" data-js="card-button">
+						Rephrase Card
+					</button>
+				</article>
+			</section>
 			<section class="classical-authors">
-				<button type="button" data-js="add-button">Add someone old!</button>
+				<h2 data-js="authors-headline">Click the button to see authors…</h2>
+				<button type="button" data-js="add-author-button">Add someone old!</button>
 				<ul class="classical-authors__list" data-js="authors-list"></ul>
 			</section>
 		</main>
