+++ sessions/js-structure/demo-end/js/Header.js
@@ -5,5 +5,5 @@ const headline = document.querySelector('[data-js="headline"]');
 
 headerButton.addEventListener('click', () => {
 	headline.textContent = `Classical ${getPlural('Author')}`;
-	headerButton.classList.add('header__button-clicked');
+	headerButton.setAttribute('disabled', '');
 });
