+++ sessions/js-structure/demo-start/css/components/card.css
@@ -1,26 +1,18 @@
 .card {
-	width: 80%;
-	margin: 0 auto;
 	display: flex;
 	flex-direction: column;
+	border: 1px solid var(--granite);
+	padding: 1rem;
+	border-radius: 8px;
 }
 
-.card + .card {
-	margin-top: 1.5rem;
-}
-
-.card-changed {
+.card--rephrased {
 	background-color: var(--water);
-	padding: 0.5rem;
 }
 
 .card__button {
 	width: 60%;
 	margin: 0 auto;
-}
-
-.card__button-clicked {
-	display: none;
 }
 
 .card__text {
