+++ sessions/js-structure/demo-start/css/components/classical-authors.css
@@ -1,23 +1,16 @@
-.classical-authors {
-	width: 80%;
-	margin: 2rem auto;
-	display: flex;
-	flex-direction: column;
-}
-
 .classical-authors__list {
 	padding: 0.5rem;
 	list-style: none;
 	display: flex;
-	flex-direction: column;
 	align-items: center;
+	gap: 1rem;
+	flex-wrap: wrap;
 }
 
 .classical-authors__list-item {
-	padding: 0.5rem;
-	width: 30%;
+	padding: 0.5rem 1rem;
 	text-align: center;
-	margin-bottom: 0.5rem;
+	border-radius: 9999px;
 }
 .classical-authors__list-item:nth-child(odd) {
 	background-color: var(--granite);
