+++ sessions/js-structure/demo-start/css/components/header.css
@@ -10,7 +10,3 @@
 	top: 10px;
 	right: 5px;
 }
-
-.header__button-clicked {
-	display: none;
-}
