+++ sessions/js-structure/demo-start/index.html
@@ -5,38 +5,45 @@
 		<meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
 		<title>JS Structure</title>
 		<link rel="stylesheet" href="./css/styles.css" />
-		<script src="./js/index.js" defer></script>
+		<script src="./js/index.js" type="module"></script>
 	</head>
 	<body>
 		<header class="header">
 			<h1 data-js="headline">JS Structure</h1>
-			<button class="header__button" type="button" data-js="header-button">Click me!</button>
+			<button class="header__button" type="button" data-js="header-button">
+				Change Headline
+			</button>
 		</header>
 		<main>
-			<article class="card" data-js="card">
-				<h2>A typical Card</h2>
-				<p class="card__text">
-					Lorem, ipsum dolor sit amet consectetur adipisicing elit. Porro veniam, mollitia
-					facilis officiis nam saepe tenetur recusandae rem corrupti labore id, delectus
-					quas quos quasi expedita pariatur esse, necessitatibus assumenda.
-				</p>
-				<button class="card__button" type="button" data-js="card-button">
-					Rephrase Card
-				</button>
-			</article>
-			<article class="card" data-js="card">
-				<h2>Another typical Card</h2>
-				<p class="card__text">
-					Lorem 2, ipsum dolor sit amet consectetur adipisicing elit. Porro veniam,
-					mollitia facilis officiis nam saepe tenetur recusandae rem corrupti labore id,
-					delectus quas quos quasi expedita pariatur esse, necessitatibus assumenda.
-				</p>
-				<button class="card__button" type="button" data-js="card-button">
-					Rephrase Card
-				</button>
-			</article>
+			<section class="cards">
+				<article class="card" data-js="card">
+					<h2>A typical Card</h2>
+					<p class="card__text">
+						Lorem, ipsum dolor sit amet consectetur adipisicing elit. Porro veniam,
+						mollitia facilis officiis nam saepe tenetur recusandae rem corrupti labore
+						id, delectus quas quos quasi expedita pariatur esse, necessitatibus
+						assumenda.
+					</p>
+					<button class="card__button" type="button" data-js="card-button">
+						Rephrase Card
+					</button>
+				</article>
+				<article class="card" data-js="card">
+					<h2>Another typical Card</h2>
+					<p class="card__text">
+						Lorem 2, ipsum dolor sit amet consectetur adipisicing elit. Porro veniam,
+						mollitia facilis officiis nam saepe tenetur recusandae rem corrupti labore
+						id, delectus quas quos quasi expedita pariatur esse, necessitatibus
+						assumenda.
+					</p>
+					<button class="card__button" type="button" data-js="card-button">
+						Rephrase Card
+					</button>
+				</article>
+			</section>
 			<section class="classical-authors">
-				<button type="button" data-js="add-button">Add someone old!</button>
+				<h2 data-js="authors-headline">Click the button to see authors…</h2>
+				<button type="button" data-js="add-author-button">Add someone old!</button>
 				<ul class="classical-authors__list" data-js="authors-list"></ul>
 			</section>
 		</main>
